Add test to make sure code field survives an edit
